fix: resolve MSVC /W4 compiler warnings across platforms
- CMakeLists.txt: define _CRT_SECURE_NO_WARNINGS for MSVC (fixes C4996)
- src/core/init.c: add explicit uint8_t casts in bapi_color_from_hex (fixes C4244)
- src/video/video.c: replace while(1) with for(;;) (fixes C4127)
- src/internal/scene/scene_internal.h: refactor bapi_hash_string (fixes C4706)
- src/platform/sdl3/platform_sdl3.c: remove redundant blend mode logic

+++ src/core/init.c
@@ -209,10 +209,10 @@ void bapi_delay(uint32_t ms)
 bapi_color_t bapi_color_from_hex(uint32_t hex_color)
 {
 	bapi_color_t color;
-	color.r = (hex_color >> 24) & 0xFF;
-	color.g = (hex_color >> 16) & 0xFF;
-	color.b = (hex_color >> 8) & 0xFF;
-	color.a = hex_color & 0xFF;
+	color.r = (uint8_t)((hex_color >> 24) & 0xFF);
+	color.g = (uint8_t)((hex_color >> 16) & 0xFF);
+	color.b = (uint8_t)((hex_color >> 8) & 0xFF);
+	color.a = (uint8_t)(hex_color & 0xFF);
 	return color;
 }
